+++ test/Action/IdentityPersistTest.php
@@ -2,27 +2,33 @@
 
 declare(strict_types=1);
 
-namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Test;
+namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Action;
 
 use Doctrine\DBAL\Connection;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\MappingKeyInterface;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Persist\IdentityPersistCreatePayload;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Persist\IdentityPersistDeletePayload;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Persist\IdentityPersistPayload;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Persist\IdentityPersistPayloadCollection;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Persist\IdentityPersistUpdatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Create\PortalNodeCreatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Create\PortalNodeCreatePayloads;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNode\PortalNodeCreateActionInterface;
-use Heptacom\HeptaConnect\Storage\Base\MappingPersistPayload;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityPersist;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNode\PortalNodeCreate;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\ContextFactory;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\EntityTypeAccessor;
-use Heptacom\HeptaConnect\Storage\ShopwareDal\MappingPersister\MappingPersister;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Repository\MappingNodeRepository;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Repository\MappingRepository;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKeyGenerator;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\Fixture\Dataset\Simple;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Test\TestCase;
 use Ramsey\Uuid\Uuid;
 use Shopware\Core\Framework\DataAbstractionLayer\DefinitionInstanceRegistry;
 
 /**
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity\IdentityPersist
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\PortalNode\PortalNodeCreate
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Content\EntityType\EntityTypeCollection
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Content\EntityType\EntityTypeDefinition
@@ -35,15 +41,14 @@ use Shopware\Core\Framework\DataAbstractionLayer\DefinitionInstanceRegistry;
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Content\PortalNode\PortalNodeDefinition
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\ContextFactory
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\EntityTypeAccessor
- * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\MappingPersister\MappingPersister
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Repository\MappingNodeRepository
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Repository\MappingRepository
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKeyGenerator
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\AbstractStorageKey
  */
-class MappingPersisterTest extends TestCase
+class IdentityPersistTest extends TestCase
 {
-    private MappingPersister $mappingPersister;
+    private IdentityPersist $identityPersistAction;
 
     private MappingNodeRepository $mappingNodeRepository;
 
@@ -64,10 +69,8 @@ class MappingPersisterTest extends TestCase
             ->getRepository('heptaconnect_mapping_node');
         $shopwareMappingRepository = $definitionInstanceRegistry
             ->getRepository('heptaconnect_mapping');
-        $shopwarePortalNodeRepository = $definitionInstanceRegistry
-            ->getRepository('heptaconnect_portal_node');
 
-        $this->mappingPersister = new MappingPersister($shopwareMappingRepository, $connection);
+        $this->identityPersistAction = new IdentityPersist($connection);
         $storageKeyGenerator = new StorageKeyGenerator();
         $contextFactory = new ContextFactory();
         $datasetEntityTypeAccessor = new EntityTypeAccessor($connection);
@@ -103,13 +106,14 @@ class MappingPersisterTest extends TestCase
         $mappingNodeKeyTarget = $this->mappingNodeRepository->create(Simple::class, $portalNodeKeyTarget);
         $this->mappingRepository->create($portalNodeKeyTarget, $mappingNodeKeyTarget, $externalIdTarget);
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->create($mappingNodeKeySource, $externalIdTarget);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistCreatePayload($mappingNodeKeySource, $externalIdTarget),
+        ]));
 
         static::assertCount(1, iterable_to_array($this->mappingRepository->listByMappingNode($mappingNodeKeySource)));
         static::assertCount(1, iterable_to_array($this->mappingRepository->listByMappingNode($mappingNodeKeyTarget)));
 
-        $this->mappingPersister->persist($payload);
+        $this->identityPersistAction->persist($payload);
 
         static::assertCount(0, iterable_to_array($this->mappingRepository->listByMappingNode($mappingNodeKeySource)));
         static::assertCount(2, iterable_to_array($this->mappingRepository->listByMappingNode($mappingNodeKeyTarget)));
@@ -130,9 +134,11 @@ class MappingPersisterTest extends TestCase
         $mappingNodeKey = $this->mappingNodeRepository->create(Simple::class, $portalNodeKeySource);
         $this->mappingRepository->create($portalNodeKeySource, $mappingNodeKey, $externalIdSource);
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->create($mappingNodeKey, $externalIdTarget);
-        $this->mappingPersister->persist($payload);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistCreatePayload($mappingNodeKey, $externalIdTarget),
+        ]));
+
+        $this->identityPersistAction->persist($payload);
 
         $targetMappings = \iterable_to_array(\iterable_filter(
             $this->mappingRepository->listByMappingNode($mappingNodeKey),
@@ -172,14 +178,15 @@ class MappingPersisterTest extends TestCase
         $this->mappingRepository->create($portalNodeKeySource, $mappingNodeKey1, $externalId1Source);
         $this->mappingRepository->create($portalNodeKeySource, $mappingNodeKey2, $externalId2Source);
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->create($mappingNodeKey1, $externalIdTarget);
-        $payload->create($mappingNodeKey2, $externalIdTarget);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistCreatePayload($mappingNodeKey1, $externalIdTarget),
+            new IdentityPersistCreatePayload($mappingNodeKey2, $externalIdTarget),
+        ]));
 
         $failed = false;
 
         try {
-            $this->mappingPersister->persist($payload);
+            $this->identityPersistAction->persist($payload);
         } catch (\Throwable $t) {
             $failed = true;
         }
@@ -222,14 +229,15 @@ class MappingPersisterTest extends TestCase
         $mappingNodeKey = $this->mappingNodeRepository->create(Simple::class, $portalNodeKeySource);
         $this->mappingRepository->create($portalNodeKeySource, $mappingNodeKey, $externalIdSource);
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->create($mappingNodeKey, $externalId1Target);
-        $payload->create($mappingNodeKey, $externalId2Target);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistCreatePayload($mappingNodeKey, $externalId1Target),
+            new IdentityPersistCreatePayload($mappingNodeKey, $externalId2Target),
+        ]));
 
         $failed = false;
 
         try {
-            $this->mappingPersister->persist($payload);
+            $this->identityPersistAction->persist($payload);
         } catch (\Throwable $t) {
             $failed = true;
         }
@@ -264,14 +272,15 @@ class MappingPersisterTest extends TestCase
         $mappingNodeKey = $this->mappingNodeRepository->create(Simple::class, $portalNodeKeySource);
         $this->mappingRepository->create($portalNodeKeySource, $mappingNodeKey, $externalIdSource);
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->create($mappingNodeKey, $externalId1Target);
-        $payload->update($mappingNodeKey, $externalId2Target);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistCreatePayload($mappingNodeKey, $externalId1Target),
+            new IdentityPersistUpdatePayload($mappingNodeKey, $externalId2Target),
+        ]));
 
         $failed = false;
 
         try {
-            $this->mappingPersister->persist($payload);
+            $this->identityPersistAction->persist($payload);
         } catch (\Throwable $t) {
             $failed = true;
         }
@@ -312,11 +321,12 @@ class MappingPersisterTest extends TestCase
         $this->mappingRepository->create($portalNodeKeyTarget, $mappingNodeKeyA, $externalIdTargetA);
         $this->mappingRepository->create($portalNodeKeyTarget, $mappingNodeKeyB, $externalIdTargetB);
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->update($mappingNodeKeyA, $externalIdTargetB);
-        $payload->update($mappingNodeKeyB, $externalIdTargetA);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistUpdatePayload($mappingNodeKeyA, $externalIdTargetB),
+            new IdentityPersistUpdatePayload($mappingNodeKeyB, $externalIdTargetA),
+        ]));
 
-        $this->mappingPersister->persist($payload);
+        $this->identityPersistAction->persist($payload);
 
         $targetMappingsA = \iterable_to_array(\iterable_filter(
             $this->mappingRepository->listByMappingNode($mappingNodeKeyA),
@@ -384,10 +394,11 @@ class MappingPersisterTest extends TestCase
                 ->equals($portalNodeKeyTarget)
         )));
 
-        $payload = new MappingPersistPayload($portalNodeKeyTarget);
-        $payload->delete($mappingNodeKey);
+        $payload = new IdentityPersistPayload($portalNodeKeyTarget, new IdentityPersistPayloadCollection([
+            new IdentityPersistDeletePayload($mappingNodeKey),
+        ]));
 
-        $this->mappingPersister->persist($payload);
+        $this->identityPersistAction->persist($payload);
 
         static::assertCount(1, \iterable_to_array(\iterable_filter(
             $this->mappingRepository->listByMappingNode($mappingNodeKey),
@@ -404,10 +415,11 @@ class MappingPersisterTest extends TestCase
                 ->equals($portalNodeKeyTarget)
         )));
 
-        $payload = new MappingPersistPayload($portalNodeKeySource);
-        $payload->delete($mappingNodeKey);
+        $payload = new IdentityPersistPayload($portalNodeKeySource, new IdentityPersistPayloadCollection([
+            new IdentityPersistDeletePayload($mappingNodeKey),
+        ]));
 
-        $this->mappingPersister->persist($payload);
+        $this->identityPersistAction->persist($payload);
 
         static::assertCount(0, \iterable_to_array(\iterable_filter(
             $this->mappingRepository->listByMappingNode($mappingNodeKey),
@@ -445,10 +457,11 @@ class MappingPersisterTest extends TestCase
                 ->equals($portalNodeKeySource)
         )));
 
-        $payload = new MappingPersistPayload($portalNodeKeySource);
-        $payload->delete($mappingNodeKey);
+        $payload = new IdentityPersistPayload($portalNodeKeySource, new IdentityPersistPayloadCollection([
+            new IdentityPersistDeletePayload($mappingNodeKey),
+        ]));
 
-        $this->mappingPersister->persist($payload);
+        $this->identityPersistAction->persist($payload);
 
         static::assertCount(0, \iterable_to_array(\iterable_filter(
             $this->mappingRepository->listByMappingNode($mappingNodeKey),
@@ -458,13 +471,14 @@ class MappingPersisterTest extends TestCase
                 ->equals($portalNodeKeySource)
         )));
 
-        $payload = new MappingPersistPayload($portalNodeKeySource);
-        $payload->delete($mappingNodeKey);
+        $payload = new IdentityPersistPayload($portalNodeKeySource, new IdentityPersistPayloadCollection([
+            new IdentityPersistDeletePayload($mappingNodeKey),
+        ]));
 
         $failed = false;
 
         try {
-            $this->mappingPersister->persist($payload);
+            $this->identityPersistAction->persist($payload);
         } catch (\Throwable $t) {
             $failed = true;
         }